--- api/hello_world.py
@@ -1,13 +1,0 @@
-from flask import Flask
-from flask_restx import Resource, Api
-
-app = Flask(__name__)
-api = Api(app)
-
-@api.route('/hello')
-class HelloWorld(Resource):
-    def get(self):
-        return {'hello': 'world'}
-
-if __name__ == '__main__':
-    app.run(debug=True)
